+++ apps/authentication/config/config.exs
@@ -9,6 +9,7 @@ config :authentication, Authentication.Guardian,
 config :authentication, Authentication.AuthenticationPipeline,
   module: Authentication.Guardian,
   error_handler: Authentication.AuthErrorHandler
+
 # This configuration is loaded before any dependency and is restricted
 # to this project. If another project depends on this project, this
 # file won't be loaded nor affect the parent project. For this reason,