+++ apps/authentication/lib/authentication/authentication_pipeline.ex
@@ -4,6 +4,6 @@ defmodule Authentication.AuthenticationPipeline do
   ## load guardian session with connection data
 
   use Guardian.Plug.Pipeline, otp_app: :authentication
-  plug Guardian.Plug.VerifySession, claims: %{"typ" => "access"}
-  plug Guardian.Plug.VerifyHeader, claims: %{"typ" => "access"}
+  plug(Guardian.Plug.VerifySession, claims: %{"typ" => "access"})
+  plug(Guardian.Plug.VerifyHeader, claims: %{"typ" => "access"})
 end